+++ Changelog.rst
@@ -8,6 +8,7 @@ New features
 ------------
 
 * Added hash functions BLAKE2b and BLAKE2s.
+* Added stream cipher ChaCha20.
 
 Resolved issues
 ---------------